# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trading Academy Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: facultyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Course list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /courses/{id}/assign-students:
    post:
      operationId: assignStudentsToCourse
      tags: [courses]
      summary: Assign students to a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignStudentsInput"
      responses:
        "200":
          description: Students assigned
  /courses/{id}/students:
    get:
      operationId: getCourseStudents
      tags: [courses]
      summary: Get students enrolled in course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  # ── Batches ───────────────────────────────────────────────────────────────
...
  /lectures:
    get:
      operationId: listLectures
      tags: [lectures]
      summary: List lectures
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lecture list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lecture"
    post:
      operationId: createLecture
      tags: [lectures]
      summary: Create recorded lecture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LectureInput"
      responses:
        "201":
          description: Created lecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
  /lectures/{id}:
    get:
      operationId: getLecture
      tags: [lectures]
      summary: Get lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
    patch:
      operationId: updateLecture
      tags: [lectures]
      summary: Update lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LectureUpdate"
      responses:
        "200":
          description: Updated lecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
    delete:
      operationId: deleteLecture
      tags: [lectures]
      summary: Delete lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Live Classes ──────────────────────────────────────────────────────────
...
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List certificates
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Certificate list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certificate"
    post:
      operationId: createCertificate
      tags: [certificates]
      summary: Issue certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Created certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
  /certificates/{id}:
    get:
      operationId: getCertificate
      tags: [certificates]
      summary: Get certificate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
    delete:
      operationId: deleteCertificate
      tags: [certificates]
      summary: Delete certificate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Notifications ─────────────────────────────────────────────────────────
...
  /device-sessions:
    get:
      operationId: listDeviceSessions
      tags: [deviceSessions]
      summary: List all device sessions
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceSession"
  /device-sessions/{id}:
    delete:
      operationId: removeDeviceSession
      tags: [deviceSessions]
      summary: Remove / force re-login for a device session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── System Settings ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, fullName, role, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        mobileNumber:
          type: ["string", "null"]
        role:
          type: string
        status:
          type: string
        profilePhoto:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        interestedCourse:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        expertise:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Course:
      type: object
      required: [id, name, description, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        thumbnail:
          type: ["string", "null"]
        status:
          type: string
        facultyId:
          type: ["integer", "null"]
        facultyName:
          type: ["string", "null"]
        studentCount:
          type: integer
        lectureCount:
          type: integer
        createdAt:
          type: string
    CourseInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        thumbnail:
          type: ["string", "null"]
        facultyId:
          type: ["integer", "null"]
        status:
          type: string
    CourseUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        thumbnail:
          type: ["string", "null"]
        facultyId:
          type: ["integer", "null"]
        status:
          type: string
    AssignStudentsInput:
      type: object
      required: [studentIds]
      properties:
        studentIds:
          type: array
          items:
            type: integer

    # ── Batch schemas ──────────────────────────────────────────────────────
...
    Lecture:
      type: object
      required: [id, title, description, courseId, youtubeVideoId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        youtubeVideoId:
          type: string
        thumbnail:
          type: ["string", "null"]
        facultyId:
          type: ["integer", "null"]
        facultyName:
          type: ["string", "null"]
        createdAt:
          type: string
    LectureInput:
      type: object
      required: [title, description, courseId, youtubeVideoId]
      properties:
        title:
          type: string
        description:
          type: string
        courseId:
          type: integer
        youtubeVideoId:
          type: string
        thumbnail:
          type: ["string", "null"]
    LectureUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        courseId:
          type: integer
        youtubeVideoId:
          type: string
        thumbnail:
          type: ["string", "null"]

    # ── Live Class schemas ─────────────────────────────────────────────────
...
    Certificate:
      type: object
      required: [id, studentId, courseId, certificateNumber, issueDate, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        certificateNumber:
          type: string
        issueDate:
          type: string
        driveFileId:
          type: ["string", "null"]
        driveViewUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    CertificateInput:
      type: object
      required: [studentId, courseId, certificateNumber, issueDate]
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        certificateNumber:
          type: string
        issueDate:
          type: string
        driveFileId:
          type: ["string", "null"]
        driveViewUrl:
          type: ["string", "null"]

    # ── Notification schemas ───────────────────────────────────────────────
...
    DeviceSession:
      type: object
      required: [id, userId, userName, userEmail, userRole, deviceFingerprint, lastLoginAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        userRole:
          type: string
        deviceFingerprint:
          type: string
        lastLoginAt:
          type: string

    # ── System Settings schemas ────────────────────────────────────────────
...